moved autogen to blueprint
