+++ api/api/app.py
@@ -19,6 +19,8 @@ from api.annotations import api_wrapper, require_login, require_teacher, require
 from api.annotations import block_before_competition, block_after_competition
 from api.annotations import log_action
 
+import api.routes.autogen
+
 log = api.logger.use(__name__)
 
 session_cookie_domain = "127.0.0.1"
@@ -26,47 +28,9 @@ session_cookie_path = "/"
 session_cookie_name = "flask"
 
 secret_key = ""
+app.register_blueprint(api.routes.autogen.blueprint, url_prefix="/api/autogen")
 
-def guess_mimetype(resource_path):
-    """
-    Guesses the mimetype of a given resource.
-
-    Args:
-        resource_path: the path to a given resource.
-    Returns:
-        The mimetype string.
-    """
-
-    mime = mimetypes.guess_type(resource_path)[0]
-
-    if mime is None:
-        return "application/octet-stream"
-
-    return mime
-
-@app.route('/api/autogen/serve/<path>')
-@require_login
-def serve_autogen_hook(path):
-    pid = request.args.get("pid", None)
-    static = request.args.get("static", "false") == "true"
-
-    tid = api.user.get_team()["tid"]
-
-    if pid not in api.problem.get_unlocked_pids(tid):
-        return WebError("You have not unlocked this problem!")
-
-    instance_number = api.autogen.get_instance_number(pid, tid)
-
-    if static:
-        instance_path = api.autogen.get_static_instance_path(pid, public=True)
-    else:
-        instance_path = api.autogen.get_instance_path(pid, instance_number, public=True)
-
-    mime = guess_mimetype(path)
-    if mime == 'text/html':
-        return send_from_directory(instance_path, path, mimetype=None, as_attachment=False, attachment_filename=None)
-    else:
-        return send_from_directory(instance_path, path, mimetype=mime)
+print(app.url_map)
 
 def config_app(*args, **kwargs):
     """
@@ -80,7 +44,6 @@ def config_app(*args, **kwargs):
     app.config["SESSION_COOKIE_NAME"] = session_cookie_name
 
     api.logger.setup_logs({"verbose": 2})
-
     return app
 
 @app.after_request
